Ubuntu_22.04: Configure system via stow (root dotfiles)
- copy root dotfiles to a new directory "RootDotfiles"
- manage them via stow

Advantage of managing the system dotfiles this way instead of using bare
git repo is that the stow packages can be installed independently of
each other, hence this is a more appropriate way of managing tweaks.

Also adding the logid.cfg file and temporary bugfix in order to disable
High resolution on MX3 master scroll wheel (otherwise the scrolling is
erratic).