PopOS: Exit with error
As the PopOS config is not up to date, we just return error if it is
executed. If PopOS will be needed we will fix bugs in the PopOS init script
and remove the exit with error command.

Also Shellcheck ignore directive was added so that Shellcheck does not
warn us about this.